add update agent password

+++ src/controllers/session_controller.js
@@ -58,12 +58,21 @@ const loadActiveSsessions = async (req, res, next) => {
     /**
      * [x] Charger les sessions actives de l'agent
      */
-    console.log(`Chargement des sessions de l'acteur`);
-    await Session.findAllByActeur(req.params.id).then(sessions => {
-        for (let index = 0; index < sessions.length; index++)
-            delete sessions[index].r_statut;
-        return response(res, 200, "Chargement terminé", sessions);
+    await Acteur.findByAgentId(req.params.id).then(async acteur => {
+        if (!acteur) return response(res, 404, `Cet agent n'existe pas !`);
+        if (req.session.e_acteur!=acteur.r_i) return response(res, 401, `L'agent ne correspond pas à l'acteur connecté !`)
+        const acteur_id = req.session.e_acteur;
+
+        console.log(`Chargement des sessions de l'acteur`);
+        await Session.findAllByActeur(acteur_id).then(sessions => {
+            for (let index = 0; index < sessions.length; index++)
+                delete sessions[index].r_statut;
+            return response(res, 200, "Chargement terminé", sessions);
+        }).catch(error => next(error));
+
     }).catch(error => next(error));
+
+
 }
 
 const destroySession = async (req, res, next) => {
@@ -71,14 +80,19 @@ const destroySession = async (req, res, next) => {
      * [x] Vérifier que la session reférencée existe
      * [x] Destruuire la session active selectionnée de l'agent
      */
-    console.log(`Destruction de la session: ${req.params.ref}`);
-    await Session.findByRef(req.params.ref).then(async session => {
-        if (!session) return response(res, 404, "Session active non trouvée");
-        await Session.destroy({
-            acteur: req.params.id, 
-            ref: req.params.ref
-        }).then(() => response(res, 200, "Session détruite"))
-        .catch(error => next(error));
+    await Acteur.findByAgentId(req.params.id).then(async acteur => {
+        if (!acteur) return response(res, 404, `Cet agent n'existe pas !`);
+        if (req.session.e_acteur!=acteur.r_i) return response(res, 401, `L'agent ne correspond pas à l'acteur connecté !`)
+        const acteur_id = req.session.e_acteur;
+        console.log(`Destruction de la session: ${req.params.ref}`);
+        await Session.findByRef(req.params.ref).then(async session => {
+            if (!session) return response(res, 404, "Session active non trouvée");
+            await Session.destroy({
+                acteur: acteur_id, 
+                ref: req.params.ref
+            }).then(() => response(res, 200, "Session détruite"))
+            .catch(error => next(error));
+        }).catch(error => next(error));
     }).catch(error => next(error));
 }
